+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - New #38: Add `setChatPermissions` method.
 - New #39: Add `exportChatInviteLink`, `createChatInviteLink`, `editChatInviteLink` and `revokeChatInviteLink` methods.
 - New #40: Add `approveChatJoinRequest` and `declineChatJoinRequest` methods.
+- New #42: Add `setChatPhoto` and `deleteChatPhoto` methods.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.